requirements.txt
#main environments 
streamlit
openai
google-generativeai 
python-dotenv 
tiktoken 

# optional dependencies for enhanced functionality
pydantic 
requests
PyPDF2